tests/requirements.txt
pytest>=7.0.0
pytest-cov>=4.0.0
requests>=2.28.0
python-dotenv>=1.0.0 